Clustering_based_method/requirements.txt

# Base ----------------------------------------
matplotlib==3.2.2
numpy==1.23.0   #numpy==1.18.5 numpy-1.23.0
opencv-python==4.5.4.60
Pillow==7.1.2
PyYAML==5.3.1
requests==2.23.0
scipy==1.10.1 #scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm==4.41.0
hdbscan==0.8.37
umap-learn


# Plotting ------------------------------------
pandas==1.1.4
seaborn==0.11.0
shapley
thop  
